requirements.txt
transformers
accelerate
peft
datasets
accelerate
evaluate
loralib
rouge-score
py7zr
trl
pyyaml